Toss SDL_types.h

+++ src/rackam/newsgroup.hpp
@@ -2,7 +2,6 @@
 #define NEWSGROUP_HPP
 
 #include<string>
-#include"SDL_types.h"
 
 using std::string;
 
@@ -12,8 +11,8 @@ class Newsgroup {
     ~Newsgroup();
 
     string name;
-    Uint64 min_message_no;
-    Uint64 max_message_no;
+    unsigned long long int min_message_no;
+    unsigned long long int max_message_no;
 
 };
