+++ src/rackam/post_file.hpp
@@ -2,7 +2,6 @@
 #define POSTFILE_HPP
 
 #include<string>
-#include"SDL_types.h"
 using std::string;
 
 #include"newsgroup.hpp"
@@ -13,8 +12,8 @@ class PostFile {
     ~PostFile();
 
     string name;
-    Uint64 max_message_no;
-    Uint64 min_message_no;
+    unsigned long long int max_message_no;
+    unsigned long long int min_message_no;
 
     Newsgroup *newsgroup;
 };
